Making motor class names not breaking
