+++ core/src/main/java/com/seattlesolvers/solverslib/solversHardware/SolversMotor.java
@@ -1,27 +1,25 @@
 package com.seattlesolvers.solverslib.solversHardware;
 
-import com.qualcomm.robotcore.hardware.DcMotorEx;
+import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.DcMotorSimple;
-
-import org.firstinspires.ftc.robotcore.external.navigation.CurrentUnit;
 
 /**
  * A wrapper DcMotor class that provides caching to avoid unnecessary setPower() calls.
  * Credit to team FTC 22105 (Runtime Terror) for the base class, we just modified it
  */
 
-public class SolversDcMotorEx {
+public class SolversMotor {
     private double lastPower = 0;
-    private final DcMotorEx motor;
+    private final DcMotor motor;
 
     private double powerThreshold = 0.0;
 
-    public SolversDcMotorEx(DcMotorEx motor, double powerThreshold) {
+    public SolversMotor(DcMotor motor, double powerThreshold) {
         this.motor = motor;
         this.powerThreshold = powerThreshold;
     }
 
-    public SolversDcMotorEx(DcMotorEx motor) {
+    public SolversMotor(DcMotor motor) {
         this.motor = motor;
     }
 
@@ -33,7 +31,7 @@ public class SolversDcMotorEx {
     }
 
     public int getPosition() {
-        return motor.getCurrentPosition();
+        return(motor.getCurrentPosition());
     }
 
     public void setDirection(DcMotorSimple.Direction direction) {
@@ -55,11 +53,11 @@ public class SolversDcMotorEx {
         return this.motor.getPowerFloat();
     }
 
-    public DcMotorEx.ZeroPowerBehavior getZeroPowerBehavior() {
+    public DcMotor.ZeroPowerBehavior getZeroPowerBehavior() {
         return this.motor.getZeroPowerBehavior();
     }
 
-    public DcMotorEx.RunMode getMode() {
+    public DcMotor.RunMode getMode() {
         return this.motor.getMode();
     }
 
@@ -67,27 +65,11 @@ public class SolversDcMotorEx {
         return this.motor.getPortNumber();
     }
 
-    public void setMode(DcMotorEx.RunMode runMode) {
+    public void setMode(DcMotor.RunMode runMode) {
         this.motor.setMode(runMode);
     }
 
-    public void setZeroPowerBehavior(DcMotorEx.ZeroPowerBehavior zeroPowerBehavior) {
+    public void setZeroPowerBehavior(DcMotor.ZeroPowerBehavior zeroPowerBehavior) {
         this.motor.setZeroPowerBehavior(zeroPowerBehavior);
-    }
-
-    public double getCurrent(CurrentUnit unit) {
-        return this.motor.getCurrent(unit);
-    }
-
-    public double getCurrentAlert(CurrentUnit unit) {
-        return this.motor.getCurrentAlert(unit);
-    }
-
-    public void setCurrentAlert(double current, CurrentUnit unit) {
-         this.motor.setCurrentAlert(current, unit);
-    }
-
-    public boolean isOverCurrent() {
-        return this.motor.isOverCurrent();
     }
 }
