+++ core/src/main/java/com/seattlesolvers/solverslib/solversHardware/SolversMotorEx.java
@@ -1,25 +1,27 @@
 package com.seattlesolvers.solverslib.solversHardware;
 
-import com.qualcomm.robotcore.hardware.DcMotor;
+import com.qualcomm.robotcore.hardware.DcMotorEx;
 import com.qualcomm.robotcore.hardware.DcMotorSimple;
+
+import org.firstinspires.ftc.robotcore.external.navigation.CurrentUnit;
 
 /**
  * A wrapper DcMotor class that provides caching to avoid unnecessary setPower() calls.
  * Credit to team FTC 22105 (Runtime Terror) for the base class, we just modified it
  */
 
-public class SolversDcMotor {
+public class SolversMotorEx {
     private double lastPower = 0;
-    private final DcMotor motor;
+    private final DcMotorEx motor;
 
     private double powerThreshold = 0.0;
 
-    public SolversDcMotor(DcMotor motor, double powerThreshold) {
+    public SolversMotorEx(DcMotorEx motor, double powerThreshold) {
         this.motor = motor;
         this.powerThreshold = powerThreshold;
     }
 
-    public SolversDcMotor(DcMotor motor) {
+    public SolversMotorEx(DcMotorEx motor) {
         this.motor = motor;
     }
 
@@ -31,7 +33,7 @@ public class SolversDcMotor {
     }
 
     public int getPosition() {
-        return(motor.getCurrentPosition());
+        return motor.getCurrentPosition();
     }
 
     public void setDirection(DcMotorSimple.Direction direction) {
@@ -53,11 +55,11 @@ public class SolversDcMotor {
         return this.motor.getPowerFloat();
     }
 
-    public DcMotor.ZeroPowerBehavior getZeroPowerBehavior() {
+    public DcMotorEx.ZeroPowerBehavior getZeroPowerBehavior() {
         return this.motor.getZeroPowerBehavior();
     }
 
-    public DcMotor.RunMode getMode() {
+    public DcMotorEx.RunMode getMode() {
         return this.motor.getMode();
     }
 
@@ -65,11 +67,27 @@ public class SolversDcMotor {
         return this.motor.getPortNumber();
     }
 
-    public void setMode(DcMotor.RunMode runMode) {
+    public void setMode(DcMotorEx.RunMode runMode) {
         this.motor.setMode(runMode);
     }
 
-    public void setZeroPowerBehavior(DcMotor.ZeroPowerBehavior zeroPowerBehavior) {
+    public void setZeroPowerBehavior(DcMotorEx.ZeroPowerBehavior zeroPowerBehavior) {
         this.motor.setZeroPowerBehavior(zeroPowerBehavior);
+    }
+
+    public double getCurrent(CurrentUnit unit) {
+        return this.motor.getCurrent(unit);
+    }
+
+    public double getCurrentAlert(CurrentUnit unit) {
+        return this.motor.getCurrentAlert(unit);
+    }
+
+    public void setCurrentAlert(double current, CurrentUnit unit) {
+         this.motor.setCurrentAlert(current, unit);
+    }
+
+    public boolean isOverCurrent() {
+        return this.motor.isOverCurrent();
     }
 }
